requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
anthropic>=0.28.0
requests>=2.31.0
pandas>=2.2.0
openpyxl>=3.1.2
reportlab>=4.2.0
python-docx>=1.1.0
boto3>=1.34.0
google-auth>=2.29.0
google-api-python-client>=2.125.0
PyGithub>=2.3.0
jira>=3.8.0
slack-sdk>=3.27.0
python-jose>=3.3.0
pyjwt>=2.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
plotly>=5.20.0
